Fix help screens.
I actually made a couple of changes to how the help screen worked, which
makes it a bit more straightforward. Normally, the help screen renders
the new last line of text (when scrolling down) to the backbuffer and
then copies the previous last n-1 lines from the frontbuffer upwards.
It then copies the line it just drew to the backbuffer to the frontbuffer.

This totally messes things up if you just want to use VW_UpdateScreen(),
as it writes to the frontbuffer directly. I've set it up to write to the
backbuffer, reordering the scroll of existing text and the writing of new
text.