Keen Dreams Steam Port open source release.
This is a release of the last version of the source code for the Steam
port of Keen Dreams I have. It was based on the keen-dreams-sdl2
project, with a number of modifications.

As-is, it's only compatible with the data files from the
now-discontinued Steam version. It also does not contain any of the code
which interacts directly with the Steam API (but there are replacement
stub versions, so you can see roughly what it did.)

Note that this is not the final version of the Steam port: it's missing
the bugfixes that were done when Nightdive Studios took over publishing
of the port. This means that there are a bunch of bugs that still exist,
and it's using the OpenGL backend, not an SDL2 backend.

This port was largely done by David Gow, with additional code by Emma
Kranz and Braden Obrzut.